Update has been completed successfully

+++ To-Let/src/main/resources/static/js/to-let.js
@@ -18,6 +18,10 @@ case 'AddServices':
 case 'View-Services':
 	$('#viewservices').addClass('active');
 	break;
+case 'Manage-View-Services':
+	$('#manageviewservices').addClass('active');
+	break;
+
 	
 default:
 	     if (menu='Home') {
@@ -41,8 +45,6 @@ setTimeout(function(){
 },1000);
 
 }
-
-
 
 
 
@@ -122,16 +124,113 @@ if(table.length){
 	  {
 		  data:'contact'
 		 
+	  }
+	  
+  ]
+	
+	
+	
+	
+});
+} 
+
+
+
+
+
+
+
+var table=$('#manageViewServices');
+
+if(table.length){
+
+	var jsonUrl=jsonurl;
+	
+
+	 table.DataTable({
+  lengthMenu:[[30,73,-1],['30 files','73 files','All']],
+  pageLength:30,
+  ajax:{
+	  url:jsonUrl,
+	  dataSrc : ''
+	 
+	  
+  },
+  columns : [
+	  
+	  {
+		data:'vCode',
+		mRender:function(data){
+			var str='';
+			
+			str+='<video src="videos/'+data+'.mp4" id="vsvideo" type="video/mp4"  controls></video>';
+			return str;
+			
+			
+			
+		}
+	  
+	  },
+		 {
+			data:'iCode',
+			mRender:function(data){
+				var str='';
+				
+				str+='<img src="images/'+data+'.jpg" id="vsimg" >';
+				return str;
+				
+				
+				
+			}
+		
+		
+	  },
+	  
+	  
+	  {
+		  data:'rentType'
+		 
+		  
+		  
+	  },
+	  {
+		  data:'country'
+		 
+	  },
+	  {
+		  data:'district'
+		 
+	  },
+	  {
+		  data:'subDistrict'
+		 
+	  },
+	  {
+		  data:'detailsAddress'
+		 
+	  },
+	  {
+		  data:'serviceDescription'
+		 
+	  },
+	  {
+		  data:'contact'
+		 
 	  },
 	  {
 		  data:'id',
+		  bSortable : false,
 		  mRender:function(data){
 			  				
 			  var str='';
-				
+					
+		
+					
 			  str+='<a href="/updateAddservice?id='+data+'" id="dbtn" class="btn btn-primary"><span class="glyphicon glyphicon-pencil"></span></a>';
+			
+			 
 			  str+='<a href="/delete?id='+data+'" id="dbtn" class="btn btn-danger"><span class="glyphicon glyphicon-trash"></span></a>';
-				 
+				
 			  return str;
 			 
 			  
@@ -147,10 +246,6 @@ if(table.length){
 	
 });
 } 
-
-
-
-
